Converted last of the javascript files over to TypeScript
Also made the files modules to be called up and easier to use.
