+++ src/delete-commands.ts
@@ -1,0 +1,21 @@
+import { REST } from '@discordjs/rest';
+import { Routes } from 'discord-api-types/v9';
+import { clientId, guildId, token } from './config.json';
+
+export default async function deleteComands() {
+  const rest = new REST({ version: '10' }).setToken(token);
+
+  // ...
+
+  // for guild-based commands
+  rest
+    .put(Routes.applicationGuildCommands(clientId, guildId), { body: [] })
+    .then(() => console.log('Successfully deleted all guild commands.'))
+    .catch(console.error);
+
+  // for global commands
+  rest
+    .put(Routes.applicationCommands(clientId), { body: [] })
+    .then(() => console.log('Successfully deleted all application commands.'))
+    .catch(console.error);
+}
